refactor(admin-registration): split single file into multiple files, refactor functions and files to be declarative, functional, less dense
