+++ core/journey/stages/stages.story.svelte
@@ -13,7 +13,7 @@
   import { initialize as initializeStyles } from '$core/style.store';
   import { buildCallbackMetadata, buildStepMetadata } from '$journey/_utilities/metadata.utilities';
   import { initCheckValidation } from './_utilities/step.utilities';
-  import AdminRegistration from './admin-registration.svelte';
+  import AdminRegistration from './admin-registration/admin-registration.svelte';
   import EmailSuspend from './email-suspend.svelte';
   import Generic from './generic.svelte';
   import Login from './login.svelte';
